backend/requirements.txt
# Core dependencies
anthropic>=0.42.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Browser automation
browser-use>=0.1.0
playwright>=1.45.0

# Async support
aiofiles>=24.1.0
asyncio>=3.4.3

# File handling
python-magic>=0.4.27
pillow>=10.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.12.0

# API utilities
httpx>=0.27.0
python-dotenv>=1.0.0

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
flake8>=7.0.0

# Optional for enhanced features
tiktoken>=0.7.0  # For token counting
rich>=13.7.0  # For better CLI output

# Deep Research Agent Dependencies
google-adk>=1.2.1  # Google Agent Development Kit
google-cloud-aiplatform>=1.60.0  # For Vertex AI
google-auth>=2.28.0  # For Google Cloud authentication
google-genai>=0.6.0  # For Gemini models 